backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
python-multipart
python-jose[cryptography]
bcrypt
authlib
itsdangerous
httpx
requests
qdrant-client
google-genai
sentence-transformers
rank-bm25
pypdf
python-docx
numpy
email-validator
